Rewriting core sim.
The core sim loop was a bit unwieldy -- it combined writing CSVs,
running roll loops, depended on internal variables, etc. Move some
code around to make it so we don't have an ever-expanding list of local
variables that we're keeping track of, in the spirit of allowing for
different types of sim to be written.